+++ src/main/java/mod/gottsch/forge/claimmyland/core/block/HorizontalAreaBlock.java
@@ -20,13 +20,22 @@
 package mod.gottsch.forge.claimmyland.core.block;
 
 import net.minecraft.core.BlockPos;
+import net.minecraft.core.Direction;
 import net.minecraft.server.level.ServerLevel;
 import net.minecraft.util.RandomSource;
+import net.minecraft.world.item.context.BlockPlaceContext;
 import net.minecraft.world.level.BlockGetter;
+import net.minecraft.world.level.LevelAccessor;
 import net.minecraft.world.level.block.Block;
 import net.minecraft.world.level.block.Blocks;
 import net.minecraft.world.level.block.SimpleWaterloggedBlock;
 import net.minecraft.world.level.block.state.BlockState;
+import net.minecraft.world.level.block.state.StateDefinition;
+import net.minecraft.world.level.block.state.properties.BlockStateProperties;
+import net.minecraft.world.level.block.state.properties.BooleanProperty;
+import net.minecraft.world.level.block.state.properties.EnumProperty;
+import net.minecraft.world.level.material.FluidState;
+import net.minecraft.world.level.material.Fluids;
 import net.minecraft.world.phys.shapes.CollisionContext;
 import net.minecraft.world.phys.shapes.VoxelShape;
 
@@ -34,10 +43,24 @@ import net.minecraft.world.phys.shapes.VoxelShape;
  * @author Mark Gottschling on 3/9/2025
  */
 public class HorizontalAreaBlock extends Block implements SimpleWaterloggedBlock {
+    // vanilla waterlogged state
+    public static final BooleanProperty WATERLOGGED = BlockStateProperties.WATERLOGGED;
+    // state to indicate whether the border is intersecting another parcel
+    public static final EnumProperty<BorderStatus> INTERSECTS = EnumProperty.create("intersects", BorderStatus.class);
+
     private static final VoxelShape SHAPE = Block.box(0, 0, 0, 16, 0, 16);
 
     public HorizontalAreaBlock(Properties properties) {
         super(properties);
+        this.registerDefaultState(this.stateDefinition.any()
+                .setValue(WATERLOGGED, Boolean.valueOf(false))
+                .setValue(INTERSECTS, BorderStatus.GOOD));
+    }
+
+    @Override
+    protected void createBlockStateDefinition(StateDefinition.Builder<Block, BlockState> builder) {
+        super.createBlockStateDefinition(builder);
+        builder.add(WATERLOGGED).add(INTERSECTS);
     }
 
     @Override
@@ -53,5 +76,26 @@ public class HorizontalAreaBlock extends Block implements SimpleWaterloggedBlock
     @Override
     public void tick(BlockState state, ServerLevel level, BlockPos pos, RandomSource random) {
         level.setBlockAndUpdate(pos, Blocks.AIR.defaultBlockState());
+    }
+
+    @Override
+    public BlockState getStateForPlacement(BlockPlaceContext context) {
+        BlockPos blockPos = context.getClickedPos();
+        FluidState fluidState = context.getLevel().getFluidState(blockPos);
+
+        return this.defaultBlockState().setValue(WATERLOGGED, Boolean.valueOf(fluidState.getType() == Fluids.WATER));
+    }
+
+    @Override
+    public BlockState updateShape(BlockState state, Direction direction, BlockState newState, LevelAccessor levelAccessor, BlockPos pos, BlockPos p_56930_) {
+        if (state.getValue(WATERLOGGED)) {
+            levelAccessor.scheduleTick(pos, Fluids.WATER, Fluids.WATER.getTickDelay(levelAccessor));
+        }
+        return super.updateShape(state, direction, newState, levelAccessor, pos, p_56930_);
+    }
+
+    @Override
+    public FluidState getFluidState(BlockState blockState) {
+        return blockState.getValue(WATERLOGGED) ? Fluids.WATER.getSource(false) : super.getFluidState(blockState);
     }
 }